requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Caching==2.1.0
Flask-Migrate==4.0.5
Werkzeug==3.0.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
PyJWT==2.8.0
Pillow==10.2.0
gunicorn==21.2.0
pytest==8.0.0
coverage==7.4.1
black==24.1.1
flake8==7.0.0 